fix bug

+++ db/migrate/20170515073304_change_product_price_column_type.rb
@@ -1,5 +1,5 @@
 class ChangeProductPriceColumnType < ActiveRecord::Migration[5.0]
   def change
-    change_column :products, :price, :decimal, precision: 10, scale: 2
+    change_column :products, :price, :float, precision: 10, scale: 2
   end
 end
