+++ db/migrate/20170515073912_change_order_total_money_column_type.rb
@@ -1,5 +1,5 @@
 class ChangeOrderTotalMoneyColumnType < ActiveRecord::Migration[5.0]
   def change
-    change_column :orders, :total, :decimal, precision: 10, scale: 2
+    change_column :orders, :total, :float, precision: 10, scale: 2
   end
 end
