+++ db/migrate/20170515074128_change_product_list_price_column_type.rb
@@ -1,5 +1,5 @@
 class ChangeProductListPriceColumnType < ActiveRecord::Migration[5.0]
   def change
-    change_column :product_lists, :product_price, :decimal, precision: 10, scale: 2
+    change_column :product_lists, :product_price, :float , precision: 10, scale: 2
   end
 end
